+++ code/system/sys.util.react/src/index.mts
@@ -1,3 +1,4 @@
-export { FC } from './FC/index.mjs';
-export { useClickOutside, useClickInside } from './hooks/useClick.mjs';
-export { useSizeObserver } from './hooks/useSizeObserver.mjs';
+export { FC } from './ui/FC/index.mjs';
+
+export { useClickOutside, useClickInside } from './ui.hooks/useClick.mjs';
+export { useSizeObserver } from './ui.hooks/useSizeObserver.mjs';
